Refactor prediction handling in main function to support modality classification and streamline output processing

+++ src/shivai/scripts/predict_multi.py
@@ -168,6 +168,11 @@ def main():
     if h5_models:
         raise NotImplementedError("Models in .h5 format are no longer supported. ")
 
+    target = meta_data['target']
+    if target == "MOD":
+        # Specific case for modality classification
+        modalities = meta_data['modalities']
+
     if keras_model:
         # Execute keras_model to have access to its classes
         # with open(keras_model) as kf:
@@ -239,25 +244,36 @@ def main():
             for j, sub in enumerate(sub_list[curr_slice]):
                 sub_pred = predictions[j].squeeze()
                 sub_pred[sub_pred < 0.001] = 0  # Threshold to remove near-zero voxels
-                subpred_im = nib.Nifti1Image(sub_pred.astype('float32'), affine=affine_dict[sub])
-                tmp_file = Path(f'tmp_{sub}_fold{fold}.nii.gz')
-                nib.save(subpred_im, tmp_file)
-                tmp_files[f'{sub}_{fold}'] = tmp_file
+                if target == "MOD":
+                    # Reusing tmp_files to directly store the predicted class here
+                    tmp_files[f'{sub}_{fold}'] = sub_pred
+                else:
+                    subpred_im = nib.Nifti1Image(sub_pred.astype('float32'), affine=affine_dict[sub])
+                    tmp_file = Path(f'tmp_{sub}_fold{fold}.nii.gz')
+                    nib.save(subpred_im, tmp_file)
+                    tmp_files[f'{sub}_{fold}'] = tmp_file
     # Taking each fold's results and averaging them
     print('Averaging the results of each model (done for each subject)...')
     for i, sub in enumerate(sub_list):
-        pred_list = [nib.load(tmp_files[f'{sub}_{fold}']).get_fdata(dtype='float32') for fold in range(len(model_files))]
-        mean_pred = np.mean(pred_list, axis=0)
-        if args.mask_files is not None:
-            brainmask = nib.load(args.mask_files[sub_list.index(sub)]).get_fdata().astype(bool)
-            mean_pred *= brainmask
-        mean_pred_im = nib.Nifti1Image(mean_pred.astype('float32'),  affine=affine_dict[sub])
         outname = args.foutname.format(sub=sub)
         if args.out_dir:
             outname = args.out_dir / outname
-        nib.save(mean_pred_im, outname)
-        for fold in range(len(model_files)):
-            tmp_files[f'{sub}_{fold}'].unlink()
+        if target == "MOD":
+            pred_list = [tmp_files[f'{sub}_{fold}'] for fold in range(len(model_files))]
+            mean_pred = np.mean(pred_list, axis=0)
+            res_dict = {mod: float(mean_pred[j]) for j, mod in enumerate(modalities)}
+            with open(outname.replace('.nii.gz', '.json'), 'w') as f:
+                json.dump(res_dict, f, indent=4)
+        else:
+            pred_list = [nib.load(tmp_files[f'{sub}_{fold}']).get_fdata(dtype='float32') for fold in range(len(model_files))]
+            mean_pred = np.mean(pred_list, axis=0)
+            if args.mask_files is not None:
+                brainmask = nib.load(args.mask_files[sub_list.index(sub)]).get_fdata().astype(bool)
+                mean_pred *= brainmask
+            mean_pred_im = nib.Nifti1Image(mean_pred.astype('float32'),  affine=affine_dict[sub])
+            nib.save(mean_pred_im, outname)
+            for fold in range(len(model_files)):
+                tmp_files[f'{sub}_{fold}'].unlink()
 
 
 if __name__ == "__main__":
